FletV2/requirements.txt
# FletV2 Requirements
flet>=0.28.3
requests>=2.32.4
psutil>=5.9.0
httpx
aiofiles>=24.1.0
python-dotenv>=1.0.0